// packages/datadog-api-client-v1/models/HostMapWidgetDefinitionStyle.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* The style to apply to the widget.
*/
export declare class HostMapWidgetDefinitionStyle {
    /**
    * Max value to use to color the map.
    */
    'fillMax'?: string;
    /**
    * Min value to use to color the map.
    */
    'fillMin'?: string;
    /**
    * Color palette to apply to the widget.
    */
    'palette'?: string;
    /**
    * Whether to flip the palette tones.
    */
    'paletteFlip'?: boolean;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): HostMapWidgetDefinitionStyle;
    static serialize(data: HostMapWidgetDefinitionStyle): {
        [key: string]: any;
    };
    constructor();
}